backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
pydantic==2.7.1
PyJWT==2.8.0
SQLAlchemy==2.0.30
mysql-connector-python==8.4.0
bcrypt==4.1.2
serverless-wsgi==3.0.3
